backend/pyproject.toml
[project]
name = "br-masters-api"
version = "0.1.0"
description = "BR Masters backend API"
requires-python = ">=3.14"

dependencies = [
    "django>=5.2,<7.0",
    "djangorestframework>=3.16,<4.0",
    "celery>=5.5,<6.0",
    "redis>=5.2,<6.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-decouple>=3.8,<4.0",
    "django-cors-headers>=4.7,<5.0",
    "gunicorn>=23.0,<24.0",
    "supabase>=2.0,<3.0",
    "pyjwt[crypto]>=2.10,<3.0",
    "httpx>=0.28,<1.0",
    "django-celery-beat>=2.7,<3.0",
    "dj-database-url>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.10,<5.0",
    "pytest-env>=1.1,<2.0",
    "ruff>=0.11,<1.0",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests/test_*.py", "tests/*/test_*.py"]
testpaths = ["tests"]

[tool.pytest_env]
DJANGO_SETTINGS_MODULE = "config.settings.test"

[tool.setuptools.packages.find]
include = ["apps*", "config*"]

[build-system]
requires = ["setuptools>=75.0,<76.0"]
build-backend = "setuptools.build_meta"
